[change]【通常ユーザ】「詳細を見る」ボタンのデザイン変更
[style.css]において，「詳細を見る」ボタンのデザインを"button_detail"として設定した．

+++ ba/models.py
@@ -75,7 +75,7 @@ class FinalScore(models.Model):
     score = models.IntegerField(default=0)
     times = models.IntegerField(default=0)  # 問題集を解いた回数
     date = models.DateTimeField(auto_now_add=True)  # 解答終了日時
-    rate = models.FloatField(default=0.0)   # 得点率
+    rate = models.IntegerField(default=0.0)   # 得点率
 
     def __str__(self):
         return f"{self.user.username}'s score for {self.question_set}:{self.times}回目 [{self.score}] 受験日{self.date}"
